+++ src/main/java/cart/dto/NewProductDto.java
@@ -1,4 +1,4 @@
-package cart.controller.dto;
+package cart.dto;
 
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
