+++ src/main/java/cart/dto/ProductDto.java
@@ -1,4 +1,4 @@
-package cart.controller.dto;
+package cart.dto;
 
 public class ProductDto {
